dsadsa

+++ src/components/List/index.jsx
@@ -17,39 +17,50 @@ export const List = ({
   activeItem,
 }) => {
   const removeList = (item) => {
-    if (window.confirm("Вы действительно хотите удалить список?")) {
-      axios.delete("http://95.163.234.208:3500/lists/" + item.id).then(() => {
-        onRemove(item.id);
-      });
+    if (item.id !== 1) {
+      if (window.confirm("Вы действительно хотите удалить список?")) {
+        axios.delete("http://95.163.234.208:3500/lists/" + item.id).then(() => {
+          onRemove(item.id);
+        });
+      }
     }
   };
 
   return (
     <ul onClick={onClick} className="list">
       {items.map((item, index) => (
-        <li
-          key={index}
-          className={classNames(item.className, {
-            active: item.active
-              ? item.active
-              : activeItem && activeItem.id === item.id,
-          })}
-          onClick={onClickItem ? () => onClickItem(item) : null}
-        >
-          <i>{item.icon ? item.icon : <Badge color={item.color.name} />}</i>
-          <span>
-            {item.name}
-            {item.tasks && ` (${item.tasks.length})`}
-          </span>
-          {isRemovable && (
-            <img
-              className="list__remove-icon"
-              src={removeSvg}
-              alt="Remove icon"
-              onClick={() => removeList(item)}
-            />
+        <div>
+          {item.id !== 1 && (
+            <div>
+              <li
+                key={index}
+                className={classNames(item.className, {
+                  active: item.active
+                    ? item.active
+                    : activeItem && activeItem.id === item.id,
+                })}
+                onClick={onClickItem ? () => onClickItem(item) : null}
+              >
+                <i>
+                  {item.icon ? item.icon : <Badge color={item.color.name} />}
+                </i>
+                <span>
+                  {item.name}
+                  {item.tasks && ` (${item.tasks.length})`}
+                </span>
+
+                {isRemovable && (
+                  <img
+                    className="list__remove-icon"
+                    src={removeSvg}
+                    alt="Remove icon"
+                    onClick={() => removeList(item)}
+                  />
+                )}
+              </li>
+            </div>
           )}
-        </li>
+        </div>
       ))}
     </ul>
   );
